+++ MorphoDig/Qt/mqMorphoDigCore.h
@@ -555,8 +555,8 @@ public:
 	void Setmui_ActiveTagMapAndRender(QString name, int numtags, std::vector<std::string> tagnames, vtkSmartPointer<vtkLookupTable> tagMap);
 	void Setmui_ActiveTag(int activeTag);
 	int Getmui_ActiveTag();
-	double Getmui_PencilSize();
-	void Setmui_PencilSize(double pencilSize);
+	int Getmui_PencilSize();
+	void Setmui_PencilSize(int pencilSize);
 
 	double* Getmui_BackGroundColor2();
 	void Getmui_BackGroundColor2(double bg[3]);
@@ -814,7 +814,7 @@ protected:
 	ActiveTagMap *mui_ActiveTagMap;
 	ExistingTagMaps *mui_ExistingTagMaps;
 	int mui_ActiveTag;
-	double mui_PencilSize;
+	int mui_PencilSize;
 
 	QString mui_LastUsedDir;
 	int mui_MoveMode;
